Functionality updates
Added Creation ability to add to inventory, employees, and admins as
well as Deletion from inventory and employees and basic exception
handling for doing so. Made specific inventory views to distinguish
between admin and employee views. Updated user portals (admin, employee,
patient, and donor) to start to reflect the functionalities that are
completed or near complete. Updated database tables for login and
validation purposes. Created a donor2 table for temporary use until
design is final.